requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.28.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
groq>=0.4.0
requests>=2.31.0
nibabel>=5.0.0
scikit-image>=0.21.0
